requirements.txt
tensorflow
numpy
scikit-learn
matplotlib
streamlit
pillow